Add linting commands (#1)
This commit adds a new lint command to the Makefile. The lint command runs flake8, pydocstyle and black to check the code and also checks the package metadata using setup.py check.

The GitHub Actions workflow is adjusted so that the linting checks also run as part of the build pipeline.